server/pyproject.toml
[tool.poetry]
name = "lucent-api"
version = "0.1.0"
description = "RESTful API using Flask"
authors = ["Jacob Brown <jdb897@uowmail.edu.au>"]
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.40.12"
pandas = "^2.0.0"
psycopg2-binary = "^2.9.0"
openpyxl = "^3.1.0"
flask = "^3.1.1"
flask-cors = "^6.0.1"
langchain-openai = "^0.3.30"
langchain-core = "^0.3.74"
langgraph = "^0.6.5"
sqlalchemy = "^2.0.43"
langchain-postgres = "^0.0.15"
langchain-community = "^0.3.27"
deap = "^1.4.3"
scikit-learn = "^1.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
